feat: add ChildFeaturePanel, ParentFeaturePanel, ProfileCard, and MatchingCandidatesNotification components
- Implemented ChildFeaturePanel for child users with verification status and feature access.
- Created ParentFeaturePanel for parent users, including subscription management and feature access based on verification and subscription status.
- Developed ProfileCard to display user profile information, including name, birth date, gender, and bio, with fallback for missing data.
- Added MatchingCandidatesNotification to inform users about matching candidates and required information for matching.
